+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.2.8
+
+- Removed `linkMode` - this extension always need list of all workspace files to work.
+- Added config `excludeScanFolders` and `cacheWorkspaceFiles`
+- Added php-parser (removed regex and line by line loop)
+- Added experimental cache workspace files list
+
 ## 0.2.0
 
 - Scaning all workspace files for match